Fix axis formatting
