+++ fiberplane-charts/package.json
@@ -13,7 +13,6 @@
     "license": "MIT OR Apache-2.0",
     "type": "module",
     "dependencies": {
-        "@visx/axis": "^3.1.0",
         "@visx/event": "^3.0.1",
         "@visx/grid": "^3.0.1",
         "@visx/scale": "^3.0.0",
