+++ fiberplane-charts/src/utils/index.ts
@@ -1,7 +1,6 @@
 export * from "./compact";
 export * from "./convert";
 export * from "./findUniqueKeys";
-export * from "./formatTimeseries";
 export * from "./getShapeListColor";
 export * from "./identity";
 export * from "./mergeRefs";
